third_party/WebKit/LayoutTests/svg/animations/animate-path-animation-cC-sS-inverse-expected.txt

SVG 1.1 dynamic animation tests
Test path animation where coordinate modes of start and end differ. You should see PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.getAttribute('d') is "M -20 -20 c 0 40 0 40 40 40 s 40 0 0 -40 Z"
PASS path.getAttribute('d') is "M -20 -20 c 10 30 10 30 40 40 s 30 5 -10 -30 Z"
PASS path.getAttribute('d') is "M -20 -20 C 10 -10 10 -10 20 20 S 30 35 -10 10 Z"
PASS path.getAttribute('d') is "M -20 -20 C 19.99 -19.99 19.99 -19.99 20 20 S 20.01 39.995 -19.99 19.99 Z"
PASS path.getAttribute('d') is "M -20 -20 c 0 40 0 40 40 40 s 40 0 0 -40 Z"
PASS successfullyParsed is true

TEST COMPLETE
